+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog].
 ### Added
 
 ### Changed
+- Reimplemented Animation Mapping System Completely `#168`
+  - This should fixes problem with objects/components at same place.
 
 ### Deprecated